完成了前端多个组件，完成挡板的事件绑定，完成GameMap的重构，完成websocket与physics模块的连接

+++ backend/src/main/java/com/example/backend/physics/objs/GizmoBall.java
@@ -20,10 +20,10 @@ public class GizmoBall extends WorldObjects{
     Color color = Color.cyan;
 
     //在我的理解中，反射得到的构造函数是这个
-    public GizmoBall(Integer id, Integer x, Integer y){
+    public GizmoBall(Integer id, Float x, Float y){
         this.objectId = id;
-        this.worldX = x.floatValue();
-        this.worldY = y.floatValue();
+        this.worldX = x;
+        this.worldY = y;
         this.r = 1;
         initBallInWorld();
     }
